+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 -------------------------
 * feat: added script to fix broken python scripts
 * fix: normalize Remote Applications feedback handling and conversions
+* fix: bulletproofed Feedback's content
 
 100.2.10 (2026-06-04)
 ---------------------